+++ backend/migrations/000028_technique_function.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE techniques
+    DROP COLUMN IF EXISTS function;